PUSHED [9 mai 2022 à 11:04]

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/10 13:34:23 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/03 21:59:33 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/08 22:14:33 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ void		ft_putstr(char *s);
 void		ft_putstr_fd(char *s, int fd);
 char		*ft_realloc(void *previous, char *new);
 char		**ft_split(char const *s, char c);
-float		ft_sqrt(float number);
+double		ft_sqrt(double number);
 char		*ft_strchr(const char *s, int c);
 char		*ft_strdel(char **adr_str);
 char		*ft_strdup(const char *s1);
